# collection.openapi.yaml
openapi: 3.0.3
info:
  title: FlyNext
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost
paths:
...
  /api/account/[uid]:
    patch:
      tags:
        - Ezio
        - account
      summary: Update Profile
      description: >-
        Primarily generated by Postbot:


        ### Update Account Details


        This endpoint allows you to update the account details for a specific
        user.


        - `uid` (path parameter) : The uid of the user.
            

        #### Request Body Parameters


        - `profilePic` (file): The profile picture of the user.
            
        - `firstName` (text): The first name of the user.
            
        - `lastName` (text): The last name of the user.
            
        - `email` (text): The email address of the user.
            
        - `phoneNumber` (text): The phone number of the user.
            

        #### Response


        Upon successful update, the endpoint returns a status code of 200 along
        with a JSON response containing:


        - `message`: A message indicating the success of the update.
            
        - `tokenUpdates`: A new access token and a new refresh token if the
        request only provides a valid refresh token.
      operationId: updateProfile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: email1@dom1.com
                firstName:
                  type: string
                  example: firstName1
                lastName:
                  type: string
                  example: lastName1
                phoneNumber:
                  type: string
                  example: '9999999999'
                profilePic:
                  type: string
                  format: binary
      responses:
        '200':
          description: ''
      security:
        - jwtBearerAuth: []
  /api/account/register:
    post:
      tags:
        - Ezio
        - account
      summary: Create Account
      description: >-
        Primarily generated by Postbot:


        The endpoint allows users to register a new account by providing their
        information. The email and the password are required.


        ### Request Body


        - `email` (string): The email address of the user.
            
        - `password` (string): The password for the new account.
            
        - `firstName` (string): The first name of the user.
            
        - `lastName` (string): The last name of the user.
            
        - `phoneNumber` (string): The phone number of the user.
            

        ### Response


        The response is in JSON format and includes the following fields:


        - `uid` (string): The unique identifier for the newly registered
        account.
            
        - `tokens` (object): An object containing the refresh token and access
        token for the registered account.
            
            - `refreshToken` (string): The refresh token for the registered account.
                
            - `accessToken` (string): The access token for the registered account.
                

        ### JSON Schema


        ``` json

        {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string"
            },
            "tokens": {
              "type": "object",
              "properties": {
                "refreshToken": {
                  "type": "string"
                },
                "accessToken": {
                  "type": "string"
                }
              }
            }
          }
        }

         ```
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                registration:
                  type: object
                  properties:
                    email:
                      type: string
                      example: email1.@example.com
                    firstName:
                      type: string
                      example: firstName1
                    lastName:
                      type: string
                      example: lastName
                    password:
                      type: string
                      example: password1
                    phoneNumber:
                      type: string
                      example: '111'
            examples:
              Create Account:
                value:
                  registration:
                    email: email1.@example.com
                    firstName: firstName1
                    lastName: lastName
                    password: password1
                    phoneNumber: '111'
      responses:
        '200':
          description: ''
...
  /api/hotel:
    post:
      tags:
        - Ezio
        - hotel
      summary: Create Hotel
      description: >-
        Primarily generated by Postbot:


        ### Add Hotel


        This endpoint allows you to add a new hotel with the provided details.


        #### Request Body


        - `logo` (file): The logo of the hotel.
            
        - `name` (text): The name of the hotel.
            
        - `address` (text): The address of the hotel.
            
        - `city` (text): The city where the hotel is located.
            
        - `starRating` (text): The star rating of the hotel.
            
        - `hotelImg` (file): An image of the hotel (up to 9).
            

        #### Response


        Upon successful execution, the server will respond with a status code of
        200 and a JSON object containing the following fields:


        - `message`: A message indicating the status of the operation.
            
        - `imgSaveFailed`: A boolean value indicating whether the images are all
        saved.
            
        - `hid`: The hid of the newly added hotel.
            
        - `tokenUpdates`: A new access token and a new refresh token if the
        request only provides a valid refresh token.
      operationId: createHotel
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                address:
                  type: string
                  example: addr1
                city:
                  type: string
                  example: city1
                hotelImg:
                  type: string
                  format: binary
                logo:
                  type: string
                  format: binary
                name:
                  type: string
                  example: hotel1
                starRating:
                  type: string
                  example: '1'
      responses:
        '200':
          description: ''
      security:
        - jwtBearerAuth: []
  /api/booking:
    get:
      tags:
        - Ezio
        - booking1
      summary: Get Bookings
      description: >-
        Primarily generated by Postbot:


        This endpoint makes an HTTP GET request to retrieve all booking
        information of a user (given the associated token) from the server.


        ### Request


        No request body parameters are required for this endpoint.


        ### Response


        - Status: 200
            
        - Content-Type: application/json
            

        The response body includes the following fields:


        - `bookings`: An array of booking information.
            
        - `itineraries`: An array of itinerary details.
            
        - `afsFailed`: A boolean value indicating if there was a failure linking
        the AFS system.
            
        - `tokenUpdates`: A new access token and a new refresh token if the
        request only provides a valid refresh token.
      operationId: getBookings
      responses:
        '200':
          description: ''
      security:
        - jwtBearerAuth: []
  /api/booking/[id]:
    patch:
      tags:
        - Ezio
        - booking1
      summary: Cancel Booking
      description: >-
        Primarily generated by Postbot:


        ### Cancel Booking Details


        This endpoint is used to cancel a specific booking identified by its
        unique ID for the associated user.


        #### Request Body


        none


        #### Response


        - Status: 200
            
        - Content-Type: application/json
            
        - `message`: A message indicating the success of the cancellation
            
        - `tokenUpdates`: A new access token and a new refresh token if the
        request only provides a valid refresh token.
      operationId: cancelBooking
      responses:
        '200':
          description: ''
      security:
        - jwtBearerAuth: []
...
  /api/hotel_search/information:
    get:
      tags:
        - William
      summary: get the hotel information by hotelId
      description: get the hotel information by hotelId
      operationId: getTheHotelInformationByHotelid
      parameters:
        - name: hotelId
          in: query
          schema:
            type: string
            example: hotel_1
      responses:
        '200':
          description: ''
      security:
        - {}
...
components:
  securitySchemes:
    jwtBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
...